refactored code, removed whitespace, moved theme selector to separate file
