+++ src/App.js
@@ -3,16 +3,15 @@ import {Route, Routes} from 'react-router-dom';
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
 import Container from 'react-bootstrap/Container'
-import Row from 'react-bootstrap/Row';
-import Col from 'react-bootstrap/Col';
-import DropdownButton from 'react-bootstrap/DropdownButton';
-import Dropdown from 'react-bootstrap/Dropdown';
 
 import AuthProvider from './components/Authentication/AuthProvider';
 import PrivateRoute from './components/Authentication/PrivateRoute.js';
 import NavBar from './components/Interface/Navbar';
 import FontSize from './components/Settings/FontSize';
 import SpeechVoices from './components/Settings/SpeechVoices';
+import Theme from './components/Settings/Theme';
+import ThemeContext, {Themes} from './components/Settings/Theme-Context';
+import FontContext from './components/Settings/Font-Context.js';
 
 import Login from './pages/Login.js';
 import Home from './pages/Home.js';
@@ -20,8 +19,6 @@ import Post from './pages/Post.js';
 import Request from './pages/Request.js';
 import RequestCreate from './pages/RequestCreate.js';
 
-import ThemeContext, {Themes} from './components/Settings/theme-context';
-import FontContext from './components/Settings/font-context.js';
 import settingsCog from "./images/settings-cog.png"
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
@@ -35,24 +32,6 @@ function App() {
     const handleShow = () => setShow(true);
     const handleClose = () => setShow(false);
 
-    function handleTheme(color) {
-        switch (color) {
-            case 'main':
-                setTheme(Themes.main);
-                localStorage.setItem('theme', JSON.stringify(Themes.main));
-                break;
-            case 'purple':
-                setTheme(Themes.purple);
-                localStorage.setItem('theme', JSON.stringify(Themes.purple));
-                break;
-            case 'orange':
-                setTheme(Themes.orange);
-                localStorage.setItem('theme', JSON.stringify(Themes.orange));
-                break;
-            default:
-                break;
-        }
-    }
 
     useEffect(() => {
         if ((JSON.parse(localStorage.getItem('theme'))) == null) {
@@ -114,27 +93,14 @@ function App() {
                                     />
                                 </Container>
                                 <Container>
-                                    <Row>
-                                        <Col xs={6} md={4}>
-                                            <DropdownButton className='' title='Default'>
-                                                <Dropdown.Item as="button">
-                                                    <button className='yellowRedButton'
-                                                            onClick={() => handleTheme('main')}></button>
-                                                </Dropdown.Item>
-                                                <Dropdown.Item as="button">
-                                                    <button className='purpleBlackButton'
-                                                            onClick={() => handleTheme('purple')}></button>
-                                                </Dropdown.Item>
-                                                <Dropdown.Item as="button">
-                                                    <button className='orangeBlueButton'
-                                                            onClick={() => handleTheme('orange')}></button>
-                                                </Dropdown.Item>
-                                            </DropdownButton>
-                                        </Col>
-                                        <Col xs={12} md={8}>
-                                            <h5 className='color-theme-description'> Adjust Color Theme</h5>
-                                        </Col>
-                                    </Row>
+                                    <Theme
+                                        theme={theme}
+                                        setTheme={(newTheme) => {
+                                            localStorage.setItem('theme', newTheme);
+                                            setTheme(newTheme);
+                                        }
+                                        }
+                                    />
                                 </Container>
                                 <Container>
                                     <SpeechVoices/>
