+++ src/components/Authentication/PrivateRoute.js
@@ -1,4 +1,4 @@
-import useAuth from '../../hooks/useAuth'
+import useAuth from '../../hooks/UseAuth'
 import {Navigate} from 'react-router-dom';
 
 const PrivateRoute = ({children}) => {
