+++ src/components/RequestCreate/RequestCreateComp.js
@@ -5,7 +5,7 @@ import Form from "react-bootstrap/Form";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 
-import "../Interface/style.css";
+import "../Interface/Style.css";
 
 export default function RequestCreate() {
     return (
